cambios backend 18102024

+++ BackEnd/loteria-nacional/msvc-juegos/src/main/java/org/loterianacional/springcloud/msvc/juegos/models/entities/CodigoPromocional.java
@@ -3,6 +3,8 @@ package org.loterianacional.springcloud.msvc.juegos.models.entities;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
@@ -25,7 +27,7 @@ public class CodigoPromocional {
     private String codigo;
 
     @Column(name = "multiplicador")
-    private Integer multiplicador;
+    private BigDecimal multiplicador;
 
     @Column(name = "activo")
     private Boolean activo;
